introduce concept of view and model as a way to separate pieces of game
