+++ src/main/java/org/snake/models/Cell.java
