+++ src/main/java/org/snake/models/Rat.java
