+++ src/main/java/org/snake/models/SnakeNode.java
